+++ main/ble_utils.c
@@ -887,7 +887,7 @@ ble_characteristic_t *ble_device_characteristic_add(ble_service_t *service,
 }
 
 ble_characteristic_t *ble_device_characteristic_find_by_uuid(
-    ble_service_t *service, ble_uuid_t uuid, uint16_t index)
+    ble_service_t *service, ble_uuid_t uuid, uint8_t index)
 {
     ble_characteristic_t *cur;
